Redesign layout and move target BPM to settings
Refactored the app layout for improved structure and responsiveness, including a new flex-based container and footer placement. Moved the Target BPM input from the main page to the Settings page, and updated the Multiplier and Settings components for better input grouping and styling. Enhanced the Tabs navigation with a logo and separated settings tab, and made various CSS improvements for consistency and usability.
