+++ src/App.css
@@ -16,6 +16,10 @@ html {
   overflow-y: scroll;
 }
 
+html, body, #root {
+  height: 100%;
+}
+
 body {
   margin: 0;
   background-color: var(--bg-color);
@@ -27,7 +31,13 @@ body {
 .app-container {
   max-width: 1152px; /* 72rem */
   margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  min-height: 100%;
  /* ---  padding: 1rem;--- */
+}
+.app-content {
+  flex-grow: 1;
 }
 @media (min-width: 768px) {
   .app-container {
@@ -70,12 +80,12 @@ body {
 
 
 /* --- Filter & Target BPM Bars --- */
-.target-bpm-bar, .filter-bar {
+.filter-bar {
   background-color: #1F2937;
   padding: 1rem;
   border-radius: 0.75rem;
   box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -2px rgba(0, 0, 0, 0.1);
-  margin-bottom: 2rem;
+  margin-bottom: 1rem;
 }
 .target-bpm-bar {
   position: sticky;
@@ -171,7 +181,7 @@ body {
 
 /* --- Dan Courses --- */
 .dan-section {
-  margin-bottom: 2rem;
+  margin-bottom: 1rem;
 }
 .dan-header {
   font-size: 1.5rem;
