+++ src/Multiplier.css
@@ -3,6 +3,9 @@
   padding: 1.5rem;
   border-radius: 0.75rem;
   box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -2px rgba(0, 0, 0, 0.1);
+}
+
+.multiplier-inner-container {
   max-width: 500px;
   margin: 0 auto;
 }
